inner_api/requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.7
django-rest-swagger==2.2.0
djangorestframework==3.12.4
freezegun==1.1.0
gunicorn==20.1.0
httpie==2.5.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
openapi-codec==1.3.2
psycopg2==2.9.1
Pygments==2.10.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
